Getting the test cases to run.
Please use the up-to-date SpringSource Maven repositories, the maven.springframework.org ones are outdated. Beyond that the libs-* ones are collective ones, which means libs-snapshot contains releases and milestones as well. See https://github.com/SpringSource/spring-framework/wiki/SpringSource-repository-FAQ for details.

Put Spring config into the classpath to be able to refer to it in the test cases without custom build tweaks.

Actually compare the size of the collection not the collection itself. Empty the collection right before the test run instead of after as someone might have added documents in the meantime or a crashing unit test can have left the collection in stale state.